requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.2.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.0
